moved PlainInput to common components
